% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.depth.local}
\alias{impute.depth.local}
\title{Single local-depth imputation}
\usage{
impute.depth.local(X, par.loc = 0.5, depth = "Mahalanobis",
  parMcd.impute = 0.5, max.iter = 25, eps.precision = 0.001,
  core.count = 1, cl.type = "SOCK", verbosity = 1, hist = FALSE)
}
\arguments{
\item{X}{Data set with missing entries.}

\item{par.loc}{Localisation parameter for the local depth.}

\item{depth}{A notion of depth to be pluged-in in the local depth chosen
among "Mahalanobis", "zonoid" (default), "Tukey".}

\item{parMcd.impute}{Parameter of the Minimum Covariance Determinan estimator
used for calculating covariance matrix required for imputation with the
Mahalanobis or spatial depth.}

\item{max.iter}{Maximum number of imputation iteraions, 25 by default.}

\item{eps.precision}{Numerical precision, i.e. maximum deviation from
previous iteration to continue, \code{0.001} by default.}

\item{core.count}{Number of cores to be used, if \code{1} (default) then no
parallelisation involved.}

\item{cl.type}{Type of the cluster, \code{"SOCK"} by default.}

\item{verbosity}{Verbosity parameter: \code{0} for no messages at
all, \code{1} (default) for minimal messages, \code{2} for a diagnostic mode
detailing each imputation iteration.}

\item{hist}{Logical indicating whether the history of each iteration should
be returned, \code{FALSE} by default.}
}
\value{
Imputed data set.
}
\description{
Imputes missing values using local depth with a pluged-in specified depth
function.
}
