% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.depth.Mahalanobis}
\alias{impute.depth.Mahalanobis}
\title{Single imputation by Mahalanobis depth}
\usage{
impute.depth.Mahalanobis(X, num.iter = 100, X.start = NULL, alpha = 1,
  mu = NULL, Sigma = NULL)
}
\arguments{
\item{X}{Data set with missing entries.}

\item{num.iter}{Number of imputation iteraions, 100 by default.}

\item{X.start}{Starting data set; if \code{NULL}, slightly perturbed mean
imputation based on existing entries serves as a starting point.}

\item{alpha}{Parameter of the Minimum Covariance Determinan estimator
used for calculating covariance matrix required for imputation with the
Mahalanobis.}

\item{mu}{Distributinal mean; if specified together with \code{Sigma}, one
imputing iteration is performed only.}

\item{Sigma}{Distributinal covariance matrix; if specified together with
\code{mu}, one imputing iteratin is performed only.}
}
\value{
Imputed data set.
}
\description{
Imputes missing values using Mahalanobis depth function.
}
