% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.depth}
\alias{impute.depth}
\title{Single imputation by data depth}
\usage{
impute.depth(X, depth = "zonoid", parMcd.impute = 1,
  depth.outsiders = "spatial", parMcd.outsiders = 1, max.iter = 50,
  max.iter.outsiders = 5, n.proj = 1000, p.extreme = 0.1,
  eps.precision = 0.001, core.count = 1, cl.type = c("SOCK", "MPI"),
  verbosity = 1, hist = FALSE)
}
\arguments{
\item{X}{Data set with missing entries.}

\item{depth}{A notion of depth to be used for imputation chosen among
"Mahalanobis", "zonoid" (default), "Tukey", "spatial", "projection",
"randomTukey", "simlicial".}

\item{parMcd.impute}{Parameter of the Minimum Covariance Determinan estimator
used for calculating covariance matrix required for imputation with the
Mahalanobis or spatial depth.}

\item{depth.outsiders}{A notion of depth to be used for imputation of
outsiders, i.e. (here) points lying on the convex hull of the data set,
chosen among "extremity" (default), "Mahalanobis", "spatial", "projection".}

\item{parMcd.outsiders}{Parameter of the Minimum Covariance Determinan
estimator used for calculating covariance matrix required for imputation
of the outsiders with the Mahalanobis or spatial depth.}

\item{max.iter}{Maximum number of imputation iteraions, 50 by default.}

\item{max.iter.outsiders}{Maximum number of outsider-imputing iterations,
5 by default.}

\item{n.proj}{Number of projections used to calculate (approximate)
projection and random Tukey depths if chosen for \code{depth} and (or)
\code{depth.outsiders}, \code{1000} by default.}

\item{p.extreme}{Portion of points in the tail to consider for the depth
extension by the extreme value theory, \code{0.1} by default.}

\item{eps.precision}{Numerical precision, i.e. maximum deviation from
previous iteration to continue, \code{0.001} by default.}

\item{core.count}{Number of cores to be used, if \code{1} (default) then no
parallelisation involved.}

\item{cl.type}{Type of the cluster, \code{"SOCK"} by default.}

\item{verbosity}{Verbosity parameter: \code{0} for no messages at
all, \code{1} (default) for minimal messages, \code{2} for a diagnostic mode
detailing each imputation iteration.}

\item{hist}{Logical indicating whether the history of each iteration should
be returned, \code{FALSE} by default.}
}
\value{
Imputed data set.
}
\description{
Imputes missing values using specified depth function.
}
