% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.ell.multiple}
\alias{impute.ell.multiple}
\title{Multiple imputation for elliptical family}
\usage{
impute.ell.multiple(X, m = 5, iter.burnin = 10, verbosity = 1)
}
\arguments{
\item{X}{Data set with missing entries.}

\item{m}{Number of multiply-imputed data sets.}

\item{iter.burnin}{Number of burn-in iterations.}

\item{verbosity}{Verbosity parameter: \code{0} for no messages at
all, \code{1} (default) for minimal messages, \code{2} for a diagnostic mode
detailing each imputation iteration, \code{3} for the graphical mode.}
}
\value{
A list with each element being a multiply imputed data set.
}
\description{
Depth-based multiple imputation for an elliptically symmetric distribution.
}
